Extract main.js from index.js
